# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Customer API
  version: 0.1.0
  description: |
    Customer-facing API for the e-banking platform (TRD §5). Draft for contract review:
    per development-process.md §3, backend and frontend build against this spec once it is
    agreed.

    Conventions:
    - Amounts are decimal strings, never floating-point numbers. Signed amounts are negative
      for debits and positive for credits.
    - CHF is the base currency.
    - Errors use the `Error` schema.

    Authentication and registration endpoints are added once the identity provider is chosen
    (development-process.md §3, Phase 0).
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
  /v1/accounts:
    get:
      tags: [Accounts]
      operationId: listAccounts
      summary: List the customer's accounts with real-time balances (FR-5)
      responses:
        '200':
          description: The authenticated customer's accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/accounts/{accountId}/transactions:
    get:
      tags: [Accounts]
      operationId: listAccountTransactions
      summary: Transaction history for one account, newest first (FR-6)
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - name: from
          in: query
          description: Only transactions at or after this instant.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Only transactions before this instant.
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Transactions, newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.0 / OIDC access token (TRD §5). The subject claim is the customer ID.
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No such resource for this customer. Other customers' resources also return 404.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [not_found, unauthorized, bad_request, step_up_required]
        message:
          type: string
    Currency:
      type: string
      enum: [CHF, EUR, USD, GBP]
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?\d+(\.\d+)?$'
          examples: ['1250.50', '-42.00']
        currency:
          $ref: '#/components/schemas/Currency'
    Account:
      type: object
      required: [id, type, name, iban, currency, status, bookedBalance, availableBalance]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [current, savings]
        name:
          type: string
        iban:
          type: string
          pattern: '^[A-Z]{2}\d{2}[A-Z0-9]{11,30}$'
        currency:
          $ref: '#/components/schemas/Currency'
        status:
          type: string
          enum: [active, pending_funding, frozen, closed]
        bookedBalance:
          $ref: '#/components/schemas/Money'
        availableBalance:
          description: Booked balance less debits still in flight.
          allOf:
            - $ref: '#/components/schemas/Money'
    AccountList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Account'
    Transaction:
      type: object
      required:
        - id
        - accountId
        - type
        - typeName
        - amount
        - status
        - holdReason
        - timestamp
        - referenceId
        - correctsTransactionId
        - category
      properties:
        id:
          type: string
        accountId:
          type: string
        type:
          description: Transaction type code from PRD §6.12.
          type: string
          pattern: '^TXN-\d{2}$'
        typeName:
          type: string
        amount:
          $ref: '#/components/schemas/Money'
        status:
          type: string
          enum: [pending, processing, completed, failed, reversed, on-hold]
        holdReason:
          type: [string, 'null']
          enum: [fraud_review, sanctions_screening, compliance_review, null]
        timestamp:
          type: string
          format: date-time
        referenceId:
          type: string
        correctsTransactionId:
          description: Set on reversal and replacement entries; the transaction they reverse or correct (FR-27a).
          type: [string, 'null']
        category:
          type: [string, 'null']
    TransactionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
...